feature(language)- Separate the composeables and add bn string resources
