+++ app/src/main/java/com/blueprint/composecrossdrill/ui/features/settings/screens/SettingScreen.kt
@@ -1,17 +1,11 @@
-package com.blueprint.composecrossdrill.ui.features.settings
+package com.blueprint.composecrossdrill.ui.features.settings.screens
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -26,17 +20,10 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Dialog
 import androidx.navigation.NavController
-import com.blueprint.composecrossdrill.R
-import com.blueprint.composecrossdrill.domain.model.language.LanguageItem
 import com.blueprint.composecrossdrill.ui.theme.spacing
 
 @OptIn(
@@ -86,51 +73,6 @@ fun SettingsScreen(navController: NavController) {
                 Text(text = "Change Language")
                 Text(text = "Tap Here to Change Language")
             }
-        }
-    }
-}
-
-@Composable
-fun ShowLanguageChangeDialog(onDismissRequest: () -> Unit) {
-    Dialog(onDismissRequest = {
-        onDismissRequest()
-    }) {
-        val languages: MutableList<LanguageItem> = mutableListOf()
-        languages.add(LanguageItem("en", "English", R.drawable.ic_english))
-        languages.add(LanguageItem("Bn", "Bengali", R.drawable.ic_bangla))
-
-        LazyColumn(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clip(MaterialTheme.shapes.medium)
-                .background(MaterialTheme.colorScheme.onPrimary)
-                .padding(MaterialTheme.spacing.small)
-        ) {
-            items(languages.size) { pos ->
-                LanguageCard(languages[pos])
-            }
-        }
-    }
-}
-
-@Composable
-fun LanguageCard(languageItem: LanguageItem) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(MaterialTheme.spacing.medium),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.Absolute.Left
-    ) {
-        Icon(
-            imageVector = ImageVector.vectorResource(languageItem.imageResId),
-            contentDescription = "Language Icon",
-            modifier = Modifier.padding(end = 8.dp)
-        )
-
-        Column {
-            Text(text = languageItem.fullForm)
-            Text(text = "Choose ${languageItem.fullForm} as default language")
         }
     }
 }
