+++ app/src/main/res/values-bn/strings.xml
